refact: basepath util
Also, I changed the name of the basePath. I was asked to.
I feel like this should be in an .env file.
But... ffffffffffffffffffffuckit

More like smooth brain.

My Christ, this commit is gonna be so embarassing to watch
I made an util just to name it wrong, then not use it
for the home page.

I think I was drunk when I made the commit before this one.